chore: rename commands.rs to command.rs
